Install community cookbooks also
Right now the best thing we can do is just copy them over, just like berks
vendor. Some future improvements may be to try to find the source URL for the
cookbook and clone it, and to update to a newer version of the cookbook when
update is run (this would be destructive, so should be behind an option).